AWS_Strands/Complex_Coding_Clara/requirements.txt
# Complex Coding Clara - Requirements
# Multi-agent coding system on AWS Strands

# Core Strands framework
strands-agents>=1.1.0
strands-agents-tools>=0.2.4

# AWS Bedrock integration (included in strands-agents)
boto3>=1.34.0
botocore>=1.34.0

# Tool packages
basic-open-agent-tools>=0.11.0

# Essential utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development tools (optional but recommended)
ruff>=0.1.0
mypy>=1.7.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
